VacacionesClustering/requirements.txt
Django==2.1.3
numpy==1.15.4
pip==18.1
pytz==2018.7
setuptools==40.6.2
beautifulsoup4==4.8.1
cycler==0.10.0
djangorestframework==3.11.0
joblib==0.14.1
kiwisolver==1.1.0
matplotlib==3.1.2
pandas==0.25.3
pyparsing==2.4.5
python-dateutil==2.8.1
scikit-learn==0.22
scipy==1.4.0
six==1.13.0
soupsieve==1.9.5
